updated index + code

+++ AMT_API_project/src/main/java/ch/heigvd/amt/amt_api_project/model/Fact.java
@@ -57,7 +57,7 @@ public class Fact implements Serializable {
     private Date dayDate;
     
     @Version
-    int version;
+    long version;
 
     public Fact() {
     }
